requirements.txt
pdf2image>=1.16.0
Pillow>=10.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10
numpy>=1.24.0
PyYAML>=6.0
